Exception handling - delete

+++ src/main/java/br/com/luisedu/course/services/UserService.java
@@ -2,8 +2,11 @@ package br.com.luisedu.course.services;
 
 import br.com.luisedu.course.entities.User;
 import br.com.luisedu.course.repositories.UserRepository;
+import br.com.luisedu.course.services.exceptions.DatabaseException;
 import br.com.luisedu.course.services.exceptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -31,7 +34,17 @@ public class UserService {
     }
 
     public void delete(Long id) {
-        userRepository.deleteById(id);
+        try {
+            if (userRepository.existsById(id)) {
+                userRepository.deleteById(id);
+            } else {
+                throw new ResourceNotFoundException(id);
+            }
+        } catch (EmptyResultDataAccessException e) {
+            throw new ResourceNotFoundException(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException(e.getMessage());
+        }
     }
 
     public User update(Long id, User user) {
